Support installation of relocatable package
This commit adds the ability to install relocatable package early on
in the setup process. It will try to use rsync to send the files (as
all file sending works in SCT) but will fall back to SCP if needed.
We are adding a new string parameter: `sync_reloc_package` which if
set to a valid package name, will upgrade each node to be updated with
the new package.