Debian repo support: Fix regular expression to account for multy achitectures
The regualr expression that is used to extract a version from a debian
repo didn't account for multiple architectures in the list file. This change
makes it account for additional architectures and saves the additional
architectures in another group that can be iterated throug.
The side effect of this is that the first mentioned arciteture will be
used for assembling the url for extracting the binary version.